requirements.txt
ipykernel
pandas
matplotlib
lightgbm
mlflow
python-dotenv